+++ __init__.py
@@ -2,6 +2,6 @@
 # SPDX-License-Identifier: MIT
 
 # Package metadata
-__version__ = '0.1.3'
+__version__ = '0.1.4'
 
 from . import acme_client, dns_rfc2136, f5_ltm, cli